requirements.txt
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
xgboost>=1.7.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0
tqdm>=4.65.0
joblib>=1.3.0
scipy>=1.11.0
streamlit>=1.20.0
